cloud test

+++ multicloud/analysis/clouds/cloud_analysis.py
@@ -47,8 +47,39 @@ def setup_background_subtraction():
 
     pass
 
-def run_cloud_analysis(cloud_name='', region=None, load=False):
+def plot_likelihoods(cloud):
 
+    likelihood_filename_base = \
+            '/d/bip3/ezbc/perseus/figures/likelihood/' + \
+            'perseus_likelihood_lee12_'
+
+    cloudpy.plot_likelihoods_hist(cloud=cloud,
+                      plot_axes=('widths', 'dgrs'),
+                      show=0,
+                      returnimage=False,
+                      filename=likelihood_filename_base + 'wd.png',
+                      limits=[10, 60, 0.0, 0.2],
+                      )
+    cloudpy.plot_likelihoods_hist(cloud=cloud,
+                      plot_axes=('widths', 'intercepts'),
+                      show=0,
+                      returnimage=False,
+                      filename=likelihood_filename_base + 'wi.png',
+                      limits=[10, 60, -1.0, 2.0],
+                      )
+
+def run_cloud_analysis(args,
+    cloud_name = 'perseus',
+    region = None,
+    load = False,
+    data_type = 'planck'):
+
+    if 1:
+        cloud_name = args['cloud_name']
+        region = args['region']
+        load = args['load']
+        data_type = args['data_type']
+        background_subtract = args['background_subtract']
 
     # define directory locations
     # --------------------------
@@ -68,14 +99,26 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
     # define filenames
     prop_filename = property_dir + \
        cloud_name + '_global_properties.txt'
-    av_filename = av_dir + \
-       cloud_name + '_av_planck_tau353_5arcmin.fits'
-    av_error_filename = av_dir + \
-       cloud_name + '_av_error_planck_tau353_5arcmin.fits'
     hi_filename = hi_dir + \
        cloud_name + '_hi_galfa_cube_regrid_planckres.fits'
     hi_error_filename = hi_dir + \
        cloud_name + '_hi_galfa_cube_regrid_planckres_noise.fits'
+    if cloud_name == 'perseus' and data_type == 'lee12':
+        av_filename = av_dir + \
+           cloud_name + '_av_lee12_iris_regrid_planckres.fits'
+        av_error_filename = None
+        av_error = 0.1
+        if background_subtract:
+            av_background = 0.5
+        else:
+            av_background = None
+    if cloud_name == 'perseus' and data_type == 'planck':
+        av_filename = av_dir + \
+           cloud_name + '_av_planck_tau353_5arcmin.fits'
+        av_error_filename = av_dir + \
+           cloud_name + '_av_error_planck_tau353_5arcmin.fits'
+        av_error = None,
+        av_background = 0.0
 
     # Plot args
     residual_hist_filename_base = figure_dir + 'diagnostics/residuals/' + \
@@ -84,11 +127,14 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
                                   cloud_name + '_residual_map'
     likelihood_filename_base = figure_dir + 'diagnostics/likelihoods/' + \
                                   cloud_name + '_likelihood'
+    av_bin_map_filename_base = figure_dir + 'diagnostics/maps/' + \
+                                  cloud_name + '_bin_map'
 
     plot_args = {
             'residual_hist_filename_base': residual_hist_filename_base,
             'residual_map_filename_base': residual_map_filename_base,
             'likelihood_filename_base': likelihood_filename_base,
+            'av_bin_map_filename_base' : av_bin_map_filename_base,
             }
 
     if 1:
@@ -104,9 +150,10 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
             cloud_name + '_diagnostic.txt'
 
     width_grid = np.arange(1, 75, 2*0.16667)
-    dgr_grid = np.arange(0.001, 0.4, 2e-4)
-    intercept_grid = np.arange(-2, 2, 0.1)
+    dgr_grid = np.arange(0.001, 0.4, 1e-3)
     intercept_grid = np.arange(-2, 2, 0.01)
+    intercept_grid = np.arange(-2, 2, 0.1)
+    #intercept_grid = np.arange(0, 1, 1)
 
     width_grid = width_grid
     dgr_grid = dgr_grid
@@ -128,12 +175,14 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
         cloud = cloudpy.Cloud(av_filename,
                               hi_filename,
                               av_error_filename=av_error_filename,
+                              av_error=av_error,
+                              av_background=av_background,
                               hi_error_filename=hi_error_filename,
                               cloud_prop_filename=prop_filename,
                               dgr_grid=dgr_grid,
                               intercept_grid=intercept_grid,
                               width_grid=width_grid,
-                              residual_width_scale=3.0,
+                              residual_width_scale=5.0,
                               threshold_delta_dgr=0.01,
                               hi_noise_vel_range=[90,110],
                               vel_range_diff_thres=2,
@@ -141,7 +190,7 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
                               verbose=True,
                               clobber_likelihoods=True,
                               binsize=binsize,
-                              diagnostic_filename=diagnostic_filename,
+                              #diagnostic_filename=diagnostic_filename,
                               plot_args=plot_args,
                               )
 
@@ -151,6 +200,8 @@ def run_cloud_analysis(cloud_name='', region=None, load=False):
         cloudpy.save(cloud, cloud_filename)
     else:
         cloud = cloudpy.load(cloud_filename)
+
+    return cloud
 
 def main():
 
@@ -163,34 +214,63 @@ def main():
     q = Queue(maxsize=0)
     num_threads = 10
 
-    clouds = {}
+    results = {}
     #clouds['perseus'] = run_cloud_analysis('perseus')
     #clouds['california'] = run_cloud_analysis('california')
     #clouds['taurus'] = run_cloud_analysis('taurus')
 
+    clouds = (
+              'perseus',
+              #'california',
+              #'taurus'
+              )
+
+    if 1:
+        for cloud in clouds:
+            args = {'cloud_name':cloud,
+                    'region':None,
+                    'load': 0,
+                    'data_type': 'lee12',
+                    'background_subtract': False}
+
+            results[cloud] = run_cloud_analysis(args)
+
     # Define a worker function for multiprocessing
-    def worker(cloud):
-        """thread worker function"""
-        try:
-            clouds[cloud] = run_cloud_analysis(cloud, load=load_clouds)
-        except KeyboardInterrupt:
-            raise KeyboardInterruptError()
+    if 0:
+        def worker(args):
+            """thread worker function"""
+            try:
+                results[cloud] = run_cloud_analysis(args)
+            except KeyboardInterrupt:
+                raise KeyboardInterruptError()
 
-        return clouds
+            return results
 
-    jobs = []
-    for cloud in ('perseus', 'california', 'taurus'):
-        try:
-            p = multiprocessing.Process(target=worker, args=(cloud,))
-            jobs.append(p)
-            p.start()
-        except KeyboardInterrupt():
-            p.terminate()
+        jobs = []
+        clouds = (
+                  'perseus',
+                  #'california',
+                  #'taurus'
+                  )
+        data_type = 'lee12'
 
-    #for cloud in clouds:
-    #    print cloud.props
+        for cloud in clouds:
+            args = {'cloud_name':cloud,
+                    'region':None,
+                    'load': True,
+                    'data_type': 'lee12'}
 
-    #clouds['taurus'] = run_cloud_analysis('taurus')
+            try:
+                p = multiprocessing.Process(target=worker, args=(args,))
+                jobs.append(p)
+                p.start()
+                p.join()
+            except KeyboardInterrupt():
+                p.terminate()
+
+        p.terminate()
+
+    plot_likelihoods(results['perseus'])
